Create Node.java
Node Class Description
The Node class defines a basic structure for a singly linked list node in Java. Each node contains:

An integer value representing the data stored in the node.

A next reference pointing to the next node in the linked list, or null if it's the last node.